Move JsoupParser tests to right package
