+++ src/test/java/com/brunodles/jsoupparser/JsoupParserTest/Builder.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.JsoupParserTests;
+package com.brunodles.jsoupparser.JsoupParserTest;
 
 import com.brunodles.jsoupparser.*;
 import org.junit.Before;
