+++ src/test/java/com/brunodles/jsoupparser/JsoupParserTest/WhenParseCollections.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.JsoupParserTests;
+package com.brunodles.jsoupparser.JsoupParserTest;
 
 import com.brunodles.jsoupparser.JsoupParser;
 import com.brunodles.jsoupparser.transformers.TransformToFloat;
