+++ src/test/java/com/brunodles/jsoupparser/JsoupParserTest/WhenParseNestedContent.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.JsoupParserTests;
+package com.brunodles.jsoupparser.JsoupParserTest;
 
 import com.brunodles.jsoupparser.JsoupParser;
 import com.brunodles.jsoupparser.collectors.TextCollector;
