+++ src/test/java/com/brunodles/jsoupparser/JsoupParserTest/WhenParseToElement.java
@@ -1,6 +1,5 @@
-package com.brunodles.jsoupparser.JsoupParserTests;
+package com.brunodles.jsoupparser.JsoupParserTest;
 
-import com.brunodles.jsoupparser.AnnotationInvocationHandler;
 import com.brunodles.jsoupparser.JsoupParser;
 import com.brunodles.jsoupparser.selector.Selector;
 import com.brunodles.test.ResourceUriResolver;
@@ -17,12 +16,18 @@ public class WhenParseToElement {
     private final JsoupParser jsoupParser = new JsoupParser.Builder().uriResolver(new ResourceUriResolver()).build();
 
     @Test
-    public void shouldReturnElementDirectly() {
+    public void shouldBeAbleToReturnElement() {
         SimpleModel simpleModel = jsoupParser.parseUrl("simple.html", SimpleModel.class);
         Element element = simpleModel.span123();
         assertEquals("magic", element.attr("class"));
         assertEquals("chain", element.attr("data-key"));
         assertEquals("wow", element.text());
+    }
+
+    @Test
+    public void shouldReturnElementDirectly() {
+        Element element = jsoupParser.parseUrl("simple.html", Element.class);
+        assertEquals("wow", element.select("#123").text());
     }
 
     public interface SimpleModel {
